refactor: reorganize compose layers
