+++ composeApp/src/commonMain/kotlin/io/github/arashiyama11/dncl_ide/ui/screen/CodingScreen.kt
@@ -1,4 +1,4 @@
-package io.github.arashiyama11.dncl_ide.ui
+package io.github.arashiyama11.dncl_ide.ui.screen
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -12,6 +12,7 @@ import io.github.arashiyama11.dncl_ide.domain.model.Entry
 import io.github.arashiyama11.dncl_ide.domain.model.NotebookFile
 import io.github.arashiyama11.dncl_ide.domain.model.ProgramFile
 import io.github.arashiyama11.dncl_ide.domain.repository.FileRepository
+import io.github.arashiyama11.dncl_ide.ui.layout.DnclIDEVertical
 import org.koin.compose.koinInject
 
 @Composable
@@ -27,7 +28,7 @@ fun CodingScreen(fileRepository: FileRepository = koinInject()) {
     }
     when (entry) {
         is NotebookFile -> NotebookScreen()
-        is ProgramFile -> DnclIDE()
+        is ProgramFile -> DnclIDEVertical()
         else -> {
             Text("No file selected or unsupported file type")
         }
