+++ composeApp/src/commonMain/kotlin/io/github/arashiyama11/dncl_ide/ui/screen/LicensesScreen.kt
@@ -1,4 +1,4 @@
-package io.github.arashiyama11.dncl_ide.ui
+package io.github.arashiyama11.dncl_ide.ui.screen
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
