+++ composeApp/src/commonMain/kotlin/io/github/arashiyama11/dncl_ide/ui/screen/SelectNotebookScreen.kt
@@ -1,4 +1,4 @@
-package io.github.arashiyama11.dncl_ide.ui
+package io.github.arashiyama11.dncl_ide.ui.screen
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
